Revert "Support Notones in alternation reductions that merge sets (#118109)" (#121775)
Clean automatic revert of #118109, plus adding a few additional tests.
The logic of the optimization was not complete. You could end up with a
negative set and then attempt to add values to it when merging in
subsequent alternation branches, and those branches would be interpreted
erroneously as negative rather than positive.