Fix DoesValueTypeContainGCRefs in the interpreter
There was a mistake in the implementation that prevented it from working
correctly in some cases. The return value of the getClassSize() is not a
number of slots filled in the array, but a number of GC references in
the array. So if there were e.g. three slots and there was one GC
reference in the second slot, the DoesValueTypeContainGCRefs returned
false.