Clarify structure within .CodeQL.yml (#121801)
This is a clarification of the structure within the existing YAML file. The cmake suppressions are not related to the ref suppressions and should be in their own independent section.

This is basically a code comment-equivalent change and doesn't impact how the analysis is performed or how suppressions are interpreted.