Avoid array -> span -> array in COSE ToBeSignedBuilder.
The ToBeSignedBuilder for hash and pure modes had an Append that accepts a ReadOnlySpan. For .NET Framework and .NET Standard, IncrementalHash and MemoryStream cannot operate on a ReadOnlySpan, so they were getting converted to an array again. There were a number of places where we already had an array though.

This change avoids to array to span to array conversation by introducing
a new AppendToBeSigned overload that accepts an array with an offset and
length, which avoids a ToArray().

Co-authored-by: Jeremy Barton <jbarton@microsoft.com>