Fix / disable several coreclr tests with interpreter (#121492)
This change contains the following:
* Fix cases when returning smaller than 32 bit int / uint values -
introduce new INTOP_RET variants like Mono
* Fix index out of range check case when native int is used as an index
to an array. To avoid adding an explosion of array element access IR
opcodes, it adds a range check opcode that is used only when the index
is nint.
* put back zeroing to INTOP_NEWOBJ_VT, few tests were failing because it
was missing and we've already fixed the issue we had with the zeroing
before by setting the "noCallArgs"
* Disable a test that uses pinvoke marshalling with calli