[release/10.0] Revert "Support Notones in alternation reductions that merge sets (#118109)" (#121775) (#121822)
Clean automatic revert of #118109, plus adding a few additional tests.
The logic of the optimization was not complete. You could end up with a
negative set and then attempt to add values to it when merging in
subsequent alternation branches, and those branches would be interpreted
erroneously as negative rather than positive.

**Customer Impact**
 
A customer hit this as part of their .NET 10 upgrade. It results in
incorrect regex matches in some cases. In their case, they had an
alternation like `.|\r|\n` (which is a non-idiomatic way of writing
“match anything”) and it was incorrectly being merged into the
equivalent of `[^\n]` (which is “match anything other than newline”).
 
**Regression**
 
Yes, in .NET 10. It was caused by an incomplete performance
optimization. Rather than try to patch the optimization, this just
reverts it, and we can see whether it’s practical to redo the
optimization correctly for .NET 11.
 
**Workarounds**
 
There are workarounds, all of which involve changing the regex pattern
to another way of expressing the same concept. The customer is changing
their code to avoid the problematic case.
 
**Testing**
 
Passing all functional and unit tests, including new ones added for the
regression.
 
**Risk**
 
Low. The optimization was done relatively late in .NET 10, so few folks
are likely to be depending on the performance gains, and the change
itself was very isolated. The revert is a clean automatic revert.