Save/restore contexts for custom notifiers (#121743)
With #121448 we are restoring contexts as we unwind during suspension.
However, since we delay the call to `OnCompleted`, we need to restore
the leaf contexts explicitly now.

Also change suspension so that we do not notify about context restores. This is
different from async1, but it is necessary to keep calling `OnCompleted` in the
right contexts when we get back to `RuntimeAsyncTask`. Also, skipping
notifications for contexts when we know no user code will be executed is one of
the optimizations we want to allow with runtime async.