JIT: Ensure proper alignment of async resumption data section (#121790)
We did not properly ensure the alignment of the async resumption data
blob, only the overall alignment of the entire data section.

Unify the code that inserts padding and use it for the async resumption
data too.

The alignment was also missing for basic block address sections, but we
did not have a problem here because we do not use absolute basic block
addresses for 64-bit targets. Still, add the code to ensure this would
get aligned correctly.

Fix #121779