Implement additional CEE_BOX peepholes (#121763)
This change implements CEE_BOX peepholes like JIT does in
Compiler::impBoxPatternMatch, mostly to fix issues with sequences
starting with CEE_BOX applied to ByRefLike types. While it is illegal to
box a ByRefLike types, some sequences can be replaced by constants or
nop.

There is also one extra fix for the destination type of an unbox. It was
incorrectly changed to StackTypeO by a large change in the past from the
correct StackTypeI.

This fixes a relatively large number of failing libraries tests.