Fix interpreter delegate calls with large alignment of first arg
There is a problem with delegate calls when the first target arch is
aligned to 16 bytes. One path to call the delegate removes the delegate
obj from the argument list by moving the arguments on the interpreter
stack by the size of the delegate obj slot. But in the problematic case,
the stack slot of the delegate obj is followed by an unused slot that
ensures the alignment of the first target argument that starts after it.
Moving the arguments by the 8 bytes garbles the arguments and makes that
unused slot part of the first target arg. Moreover, it also doesn't move
the last 8 bytes of the last argument due to this.

This change fixes it by ensuring that we move the arguments starting at
the aligned location of the first target argument.