JIT: Skip GC info for trivial async calls for x86
For x86 we only report GC pointers live in callee save registers, and we
also avoid reporting any trivial call.
There is compensating code upstream of the GC reporting that returns for
trivial calls, but this code did not account for the GC ref returned for
async continuations. Thus it was possible that we made it into the GC
encoder with a trivial call and then hit an assert.