Switch to JIT intrinsics for generic context/async continuations (#121799)
NativeAOT and coreclr have various places where they want to have
explicit control over the generic context and async continuation
arguments. Introduce JIT intrinsics that allow them to explicitly set
what value should be used for these arguments in an upcoming call. Model
these arguments as part of the calling convention in the signatures
where they are passed. That fixes #118575 for async, and is also
necessary to compute proper call stubs for the interpreter.

See #121781 for more context.