Implement InstantiateSignature on AsyncMethodVariant/EcmaMethod (#121948)
We left this as NYI but I hit this. The correct implementation for
anything that represents a definition should be `return this`. Since I
was looking at this, I'm doing the same for `EcmaMethod` too since the
base `MethodDesc.InstantiateSignature` does a lot of work for nothing on
a definition.

Cc @dotnet/ilc-contrib