Arm64: Initialization of TYP_SIMD/TYP_MASK locals
TYP_SIMD/TYP_MASK locals are zero-initialized by writing the entire
UnknownSizeFrame through a short loop of vector stores.

TYP_SIMD locals are poisoned individually for debugging code, by
broadcasting the poison value to a vector and storing it. TYP_MASK
locals should not be poisoned, because they should not be address
exposed.

Code for both scenarios is emitted in the first basic block of the
function.