Breaking change doc tool (#120987)
This adds a workflow to run a script when a PR is labeled as
`needs-breaking-change-doc-created`,

The script collects details about the PR, the breaking change issue
template, and sample issues then asks an LLM to generate an issue based
on that.

The issue content is then encoded into a clickable link which is then
added as a comment to the PR.

I'm doing this from a dotnet branch so that I can hopefully create a PR
off of it to demonstrate the workflow in action.