Stop spilling ldvirtftn (#120866)
This PR stops spilling the target for calls implemented via `ldvirtftn` in the
importer, which is the way we implement GVMs. We were previously spilling the
target with `CHECK_SPILL_NONE`, but this is incorrect because it induces a null
check (see #121711). So this PR fixes that issue. Also, avoiding the spill gives
us a simple uncontroversial way to devirtualize in some cases when we can
extract information directly from `gtCallAddr`.

The problem with removing the spill is code size regressions. `gtCallAddr` is
normally evaluated after arguments, and the backend is not good at handling the
ABI constraints of the arguments when `gtCallAddr` contains calls. To counteract
that we introduce a new optimize in lowering. The optimization tries to move the
evaluation of the target to happen before the arguments. It effectively
accomplishes what the previous `CHECK_SPILL_NONE` spill was doing, except with
the right legality checks.


Contributes to #112596

Fixes #121711

---------

Co-authored-by: Jakob Botsch Nielsen <jakob.botsch.nielsen@gmail.com>