Reliable clean-up of SafeHandles in Deflater/Inflater (#114826)
Fixes #89445

This reimplements #71991 in `Deflater` and `Inflater`. The PR was
previously reverted in #85001 as a result of issue #84994.

As noted in the reverting PR, `Deflater` simply needed to dispose of the
returned SafeHandle and supress its own finalizer if
`CreateZLibStreamForDeflate` threw an exception. We now do this.

We couldn't do this for `Inflater` at the time - it had to support
concatenated GZip data, and as part of that support it would recreate
its `ZLibStreamHandle`. As a result of #113587, this recreation no
longer happens and we can treat `Inflater` as we would treat `Deflater`.
I've refactored `Inflater.InflateInit` out of existence and made
`_zlibStream` readonly to reflect this.

This also incorporates a change to both class' `DeallocateBufferHandle`
methods (when called by `Dispose`) to prevent them dereferencing a
`ZLibStreamHandle` after it's been disposed of. This is roughly how
#84994 was originally discovered.

---------

Co-authored-by: Radek Zikmund <32671551+rzikm@users.noreply.github.com>