[RuntimeAsync] Skip over nonuser continuations when searching for awaiter configuration (#121976)
Fixes: #121971

Sometimes we have transparent/infrastructure code between user code and
a ValueTask-returning method that wraps IValueTaskSource. An example is
an unboxing stub.

This is the scenario where we need to figure the configuration of the
await operation by looking at continuation that corresponds to the user
code (only those are configured).
In a case if we see transparent continuations, we should just continue
walking continuation chain.