Runtime-async continuation object reading with Cordb (#121432)
This PR ensures that we are able to access continuation objects through
APIs such as CordbObjectValue::GetFieldValue. As the dynamically
generated continuation types with the local variable fields do not have
metadata, we use the parent class which does have metadata. We will
therefore not be able to view the locals fields through these APIs, but
with the low likelihood that someone is digging this deep into
implementation details, this should be an acceptable trade-off.