Refactor `GenTreeJitIntrinsic` to store the base type for SIMD operations as a `var_types`. (#121251)
This PR adjust `GenTreeJitIntrinsic` to save the base type of SIMD
operation as a `var_types` and not `CorInfoType`. This is done to allow
SIMD operations on types that are not part of the CIL.

Some notes on the changes:

1. `simdBaseJitType` has been renamed to `simdBaseType` and most flows
that previous had both a `simdBaseJitType` and `simdBaseType` now only
use the `simdBaseType`.
2. `simdBaseType` should now refer to the type returned via
`JitType2PreciseVarType`.
3. Some instances required the less precise, `JitType2VarType`.
4. `gtAuxiliaryJitType` has been left unchanged. Some new functions have
been introduced called `getCorBaseTypeXX` that keep the original
behavior.

If we decide to also change `gtAuxiliaryJitType` I can make the changes
and refactor things a bit more.

Current opening PR as draft to get a test run.

---------

Co-authored-by: Andy Ayers <andya@microsoft.com>