Build the test tree with TrimMode=full (#121697)
We're likely losing out on test coverage by not testing the recommended
mainstream shipping configuration.

The test changes are all to make sure we don't trim what's needed.

Where I could, I avoided just putting TrimMode=full on the assembly
(because that makes the test case RequiresProcessIsolation and that's
not great). For some I had to give up because there was no clean way to
do it.