JIT: Handle `FIELD_LIST` for tailcall arg/parameter interference checks
Since tailcalls place the arguments in the incoming parameter area it
corrupts the value of existing parameters to the function. To handle
this there is logic in place to create defensive copies of the
parameters if we detect that we will need those later in the tailcalling
sequence. However, this logic was missing handling when the argument was
a `FIELD_LIST`:

- When an argument was a `FIELD_LIST` we did not take into account that
  we needed to look for parameter uses starting from its operands. Fix
  this by introducing a `FirstUncontainedOperand` function and use it
  here.

- When the value of a `PUTARG_STK` was a contained `FIELD_LIST` we did
  not take into account that its codegen may invalidate the values of
  its own operands. Detect this case and introduce temps for it.