Ensure we fetch all PR detail data (#122078)
As part of the cleanup in PR
https://github.com/dotnet/runtime/pull/120987/commits/d27e210a908455c9ebf1b930e777a124f1f38789#diff-2ba828c15ad12c13f308f51998a4c9240d7c7325135ff7eaaff120e20dd61166L772

I tried to reduce the number of CLI calls when a PR number was
specified. When doing so I didn't consolidate the JSON fields, nor did I
update the source of the info later in the script.

This addresses both problems. I've validated with local runs - example -
https://github.com/dotnet/runtime/pull/121938#issuecomment-3598438340