Fix interpreter delegate args alignment
The recent fix to handle alignment of the first argument of a delegate
call when we were removing the delegate obj from the stack was not
complete. It only handled the case when the first argument required
16 byte alignment. But the same issue exists if any of the args requires
this alignment.

This change fixes it by finding the first argument that requires 16 byte
alignment and then moving the args before that as usual and moving
that arg and all following ones to 16 byte aligned location. This preserves
the alignment for the rest of the stuff too.