Fix FrameDataAllocator popping in InterpreterFrame::ExceptionUnwind (#122164)
The InterpreterFrame::ExceptionUnwind_Impl was passing a wrong topmost
frame to the frameDataAllocator.PopInfo - the one read directly from the
m_pTopInterpMethodContextFrame which is not always up to date. The
correct way is to use the call to GetTopInterpMethodContextFrame to get
it.

It was leading to the following assert in some libraries tests:
assert(pCurrent == pFirst && pCurrent->pFramePos ==
pCurrent->pFrameStart)