Use `CorSigUncompressCallingConv` to uncompress calling conventions in `PrettyPrintSignature` (#122177)
We were using `CorSigUncompressData` on the calling convention byte, but
this is not stored as a normal compressed integer. It happened to work
because compressed integers are specified by having their upper bit set,
and metadata signatures never have that bit set in their calling
convention. However, for some internal cases we repurpose that bit
(specifically to specify that a signature has a generic context
parameter).

Fix #122050