[clr-interp] Runtime async support (#121862)
This PR adds runtime async support to the interpreter

Notable changes:
1. Wrap async method bodies in a try/finally block to capture/restore
the exec and sync contexts, just like synchronized methods.
2. Add a series of instructions for doing a suspend/resume. This set is
currently somewhat longer than ideal, and we could simplify this by
having a set of managed helper functions to reduce the number of
instructions needed.
3. The continuation is filled in with a copy of the data from the stack.
4. The calling convention is non-traditional for the extra return value.
For the continuation argument, instead of representing it on the stack
as part of the return value, we represent it as a field on the
InterpreterFrame, and update and check that frame as needed. This allows
the vast majority of the code in the compiler to ignore the concept of
runtime-async.
5. The call stub generator is also modified to pass the extra parameter,
but more significantly, all of the stubs which call into the interpreter
are setup to now capture the continuation argument and put it into the
special return register.
6. Tests have been updated to do less work when running under the
interpreter. Some of the existing tests were designed specifically to
trigger various rejit thresholds, and while that is a fine thing to test
in the JIT, the interpreter is FAR slower, and needs a different
approach.

TODO before merge
~1. Remove the changes to enable RuntimeAsync by default.~
~2. Add documentation about how the interpreter async abi works~
~3. Add documentation about what a code generator needs to do to make
runtime async work~
~4. I'm pretty sure IL peep handling has a bug handling any of the 2
byte opcodes. This needs to be tested.~

---------

Co-authored-by: Jakob Botsch Nielsen <Jakob.botsch.nielsen@gmail.com>