[clr-interp] Fix issue where a delegate directly wraps another delegate In the interpreter we use the CallStub of a delegate invoke method for both invoking the Invoke method directly as well as for handling the calling convention details for a fallback call to use the function pointer embedded in the delegate object.
In the current logic when we construct the call stub for for the `Invoke` method for the use of just doing signature handling for delegate invoke, we didn't supply it with a target pointer, we just use NULL. When actually dispatching, we copy the `CallStubHeader` to the stack and then fill in the target pointer. It turns out that this is inadequate when a delegate directly wraps a delegate. Since we share the CallStubHeader for both purposes, we end up simply dispatching to `NULL` which will cause an access violation. The fix is to always fill in the CallStubHeader completely, but still stomp the target pointer when we are doing the delegate invoke path.

Included in this change is a coreclr test case to avoid the need to have some code in the xunit testharness which hits this in the libraries test.