[clr-interp] Fix Runtime_74635_1 test (#122194)
This is an interesting test which touches on a poorly design/documented
portion of the ECMA spec. Notably, what can be done with the this
pointer of a structure during its construction. Notably, if the this
pointer is stored somewhere else, and then that pointer is used then
there is a possibility of mutating a value on the IL evaluation stack in
the current interpreter implementation. As that is not considered a
thing which should be possible, RyuJIT has the policy of creating
function local temporaries for the newobj instruction, and then copying
the result to a location on the evaluation stack. This change matches
that behavior which is the current reading of the correct behavior
according to the ECMA standard.