Enable interpreter stack overflow handling (#122204)
Stack overflow of the interpreter stack was not being handled yet. The
stack overflow coreclr test was failing due to that.

This change adds checks when a called method frame would not fit the
interpreter stack and fires the stack overflow handling in that case.

I also had to fix the Thread::UnwindToFirstManagedFrame to work for the
interpreter frames. Before that, it would not detect them and it would
unwind through the InterpExecMethod and further native frames.