JIT: Disable flag optimization on arm32 again (#122215)
arm32 does not set the right flags for these operations, so it cannot be
enabled.

I enabled this in #121380, and these tests did fail there too, but I
mistakenly thought the failure was preexisting and the same as #121659.

Also do a small refactoring while I'm here, to make use of
`SupportsSettingFlagsAsCompareToZero` from `SupportsSettingZeroFlag`.

Fix #122185