Improve stack overflow stacktrace pretty print
The current algorithm used for the stack overflow stactrace pretty
printing isn't able to find repetitions that don't start at the first
frame. That leads to huge stack traces when there are few non-repeated
frames on the top of the stack, e.g. when a recursion in managed
method calls to some common chain of methods and the stack overflow
occurs down that chain and not in the recursive part.

This change fixes that. When no repeated sequence is found at the top of
the stack, it tries to search from the next frame and so on until
a repeated sequence is identified.

Close #118218