Fix TryEnsureSufficientExecutionStack with interpreter
The ReflectionInvocation::TryEnsureSufficientExecutionStack was not taking into
account the fact that with interpreter enabled, the code can be executed either
by the interpreter or it can be AOTed. The interpreter has its own stack with
its own limit.

3 libraries tests suites were failing due to this problem.

As a best effort fix, this change adds check for the interpreter stack too and
the method returns true only if there is enough space on both real and the
interpreter stack, as it is not possible to decide which of them will be
used.