Remove support for running illink on runtime tests while testing
We have not been running this mode for years. The test infra also dates back to before we had reflection analysis, so we aren't even testing the most complicated portion of the linker.

This testing is better validated through our NativeAOT testing, which runs its trimming implementation, as well as our test legs that run the trimmer on the libraries tests.