Fix Double/Single.ConvertToIntegerNative for non-primitive types
The current logic handles the
System.Double/Single.ConvertToIntegerNative always as must expand and
fails with NO_WAY if the integer type is not primitive. However, that
makes it incorrectly fail e.g. when the integer type is Int128.

This change fixes it by using the compiled version of the function for
non-primitive types.