Re-Enable JIT calls for Wasm Target in Crossgen (#122334)
This PR reverts the behavior from #122111 which stubs out a method
instead of calling the JIT for the crossgen Wasm target. Now, the JIT
will be invoked as normal so that the output of method codegen can be
placed in a simple object file.

Output code currently won't run (I tried this) because we aren't yet
emitting a correct local count at the beginning of compiled methods.