[release/10.0] JIT: Handle `FIELD_LIST` for tailcall arg/parameter interference checks (#122287)
## Customer Impact

- [X] Customer reported
- [ ] Found internally
 
The JIT may corrupt the values of arguments when forwarding parameters
to other functions as part of a tailcall. Reported by customer in
#122138.

## Regression

- [x] Yes
- [ ] No

This is a regression in .NET 10 where we started tailcalling in new
scenarios not covered by some existing handling.

## Testing

Regression test included.

## Risk

Low. Detect the case in the existing handling and introduce additional
defensive copies for it.