Disable bad address writing interop tests with interpreter (#122308)
This change disables all libraries interop tests that try to check that
access to non-null invalid address throws access violation or null
reference exception when executed using the interpreter.

I've introduced a new `PlatformDetection.SupportsDirtyAccessViolations`
for this purpose that is set for coreclr interpreter and for Mono and
WASM too. So I have also removed the pre-existing attribute on these
tests that was disabling them on Mono.