Box value types implementing `IXmlSerializable` in `XmlSerializer` generated IL
Value types implementing the `IXmlSerializable` interface were not being
property boxed when they declared a parameterless constructor. This
caused the `XmlSerializer` to fail to deserialize them with an
`InvalidProgramException` exception when the generated IL tried to call
the `XmlSerializationReader.ReadSerializable` method with the value.

Fixes [#99613](https://github.com/dotnet/runtime/issues/99613)