[clr-interp] Fix data loss issue at IL merge points (#122247)
Since the interpreter uses either 32bit or 64 bit floats for values on
the stack it is possible to be in a situation where we lose precision in
a floating point value at an IL merge point. The architecture of the
compiler doesn't have a scheme which would allow the compiler to avoid
hitting this problem, so instead, I've built a scheme to record the
problematic details, and retry the entire compilation of the method
while providing a way to carry the data that certain stack slots need to
be more precise than previously expected.

This fixes jit\jit64\opt\regress\vswhidbey\481244\foo2_d which tests for
this exact scenario

In addition, on 64bit platforms we also have some situations where we
can silently upgrade to a 64 bit int, this covers that situation too, as
well as the case where we merge a native integer and a ByRef so we
should treat the result as a ByRef to avoid any cases where we lose
track of a GC pointer.