Fix WASM exception propagation through native runtime
This change fixes propagating managed exceptions into native frames
on WASM. It ensures that the exception is ignored in the
`UNINSTALL_MANAGED_EXCEPTION_DISPATCHER_EX` until it reaches the
target frame. This is necessary because we cannot unwind the
stack up to the target frame on WASM unlike other platforms.