[release/9.0-staging] Backport Reliability fixes for the Thread Pool (#122362)
Fixes: #121608
Backport of https://github.com/dotnet/runtime/pull/121887

The ThreadPool in rare cases allows a scenario when an enqueued workitem
is not guaranteed to be executed unless more workitems are enqueued. In
some scenarios execution of a particular workitem may be necessary
before more work is enqueued, thus leading to a deadlock.
This is a subtle regression introduced by a change in enqueuer/worker
handshake algorithm.

The same pattern is used in 2 other ThreadPool-like internal features in
addition to the ThreadPool.

## Customer Impact

- [ ] Customer reported
- [x] Found internally

## Regression

- [x] Yes
- [ ] No

## Testing

Standard test pass for deterministic regressions.
A targeted stress application that demonstrates the issue. 
(no fix: hangs within 1-2 minutes, with the fix: runs for 20+ min. on
the same system)

## Risk
Low. This is a revert to the preexisting algorithm for the
enqueuer/worker handshake. (in all 3 places)