[Wasm RyuJit] Fixes for some issues in codegen bringup tests (#122483)
The strategy for marking blocks that need labels was wrong. We need to
mark block ends when we encounter block starts. To do this efficiently
we need to map from interval end index to block, so expose the map we
built during interval building for use in codegen.

We likely don't need Just My Code debug support for Wasm, so ignore any
request from the runtime to insert the checks.

Wasm structured control flow doesn't allow us to keep blocks in
increasing IL offset order (in general). So disable some of the scope
tracking mechanisms that expect to see this.

Fix another instance where we assumed a 32 bit target would decompose
longs.