Don't runtime-await Delegate.Invoke methods (#122491)
This is doubly inefficient because:

* The method pointed to by the delegate is guaranteed not runtime-async
(so we'd go through up to two thunks), and
* the variant of `Invoke` method is no longer _the_ `Delegate.Invoke`
method that can be reported as `CORINFO_FLG_DELEGATE_INVOKE` to RyuJIT
(needs to be invoked as a regular method instead of the efficient
delegate invoke sequence).

CoreCLR does this in methodtablebuilder.cs where we consider all methods
on delegates as `MethodReturnKind::NormalMethod`.

Also adding a test because it looks like we don't have one.