Fix scenarios involving GVM on reflection-blocked methods (#122497)
GVM dispatch didn't work when calling an async GVM from a dynamic
context (e.g. we are in a GVM and trying to call another GVM).

"A generic virtual method was called" is synonymous with "we need a
`RuntimeMethodHandle` of a GVM" in the compiler since the GVM dispatch
sequence uses the `RuntimeMethodHandle` to identify the GVM slot.

Whenever we generate a `RuntimeMethodHandle` (either the static one, or
the one in native layout used when runtime type loading) we need to
consider it a GVM call. We were correctly considering it in the static
case
(https://github.com/dotnet/runtime/blob/bed0cadfbf6a7067e24e302549829a45475cd224/src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/DependencyAnalysis/RuntimeMethodHandleNode.cs#L57),
however we missed the dynamic case (native layout information for
`RuntimeMethodHandle` I'm updating here).

We were getting away with this because the line above the one I'm
updating (`factory.MetadataManager.GetDependenciesDueToLdToken`) would
drop a `ReflectedMethodNode` into the graph that also means "GVM method
is called". However we will not get a `ReflectedMethodNode` for
something that is reflection-blocked, such as async variants. So we need
to explicitly spell out the dependency so that this works for runtime
async.