Fix dataflow analysis of lambdas in async methods (#122450)
When we enter ILLinker part of the ILC codebase, we need to switch to
the mode where async variants don't exist and async methods have
mismatched bodies. There was a spot where we didn't do the switch here:


https://github.com/dotnet/runtime/blob/d31e5990b896447bfc3dbe98cfe6ec3b169a4896/src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/UsageBasedMetadataManager.cs#L939-L949

The fix is to do the switch early in `FlowAnnotations` so that the
`ILProvider` `FlowAnnotations` use (and other parts take advantage of)
is already ILLinkerified.

Fixes the situation shown in the regression test (we were analyzing the
IL of the async variant of `RuntimeAsyncWithLambda` and missed analyzing
the lambda completely).