[wasm][coreclr] Fix DoPrestub for delegates
Return early for delegates to avoid setting the slot in
`SetCodeEntryPoint` and to skip backpatch. The delegates are special
case, where we have single implementation for all of them, and we
don't want to break other parts, which depend on 1:1 mapping between
implementation and method desc.

This fixes https://github.com/dotnet/runtime/issues/121955, which was
still happening a lot in the runtime tests (cca 90 occurences in the
out of process tests on wasm)

We were getting the exception, because during the access checks we got
private CreateDelegate method from the MT slot.