[mono][interp] Disable inlining of methods that always throw (#122587)
When inlining, we create a placeholder exit_bb that is branched to
whenever the inlined method returns. When this bblock was not reachable,
it was leading to some inconsistent CFG state. This is relatively easy
to fix but, if exit_bb is dead, it means that the method is always
throwing an exception. We shouldn't bother inlining such methods in the
first place, since they are not part of hot code paths.

Fixes https://github.com/dotnet/runtime/issues/122529