Insert a set of known types and methods references necessary for async thunks into the mutable module (#121679)
Async thunk IL has references to methods and types that may not have MethodRef
or TypeRef tokens in the assembly being crossgen'ed. There are some
assertions to validate that any references in "faux il" version with the
assembly, which isn't true for these types and methods. To work around
this, we can inject these references into the mutable module, a
synthetic module added to the assembly for things like this. When we are about
to compile an async thunk method, we check to see if all the required
references are already in the mutable module, and add them if not.