[RyuJit/WASM] Register allocator skeleton (#122414)
Force all locals to stack and add an interface for the SP/FP.

Implement `GT_LCL_ADDR` / `GT_LCL_FLD` as samples.

Next step is to add the stackifier (it'll run before RA and liveness,
since it may need to introduce locals and we want to be able to
enregister them). After that - parameter homing, and that should bring
us to syntactically (if not yet semantically) valid WASM.

---------

Co-authored-by: Andy Ayers <andya@microsoft.com>